+++ orm/src/main/java/org/hibernate/tool/internal/export/lint/HbmLint.java
@@ -1,14 +1,15 @@
-package org.hibernate.tool.hbmlint;
+package org.hibernate.tool.internal.export.lint;
 
 import java.util.ArrayList;
 import java.util.List;
 
 import org.hibernate.boot.Metadata;
+import org.hibernate.tool.hbmlint.Detector;
+import org.hibernate.tool.hbmlint.Issue;
 import org.hibernate.tool.hbmlint.detector.BadCachingDetector;
 import org.hibernate.tool.hbmlint.detector.InstrumentationDetector;
 import org.hibernate.tool.hbmlint.detector.SchemaByMetaDataDetector;
 import org.hibernate.tool.hbmlint.detector.ShadowedIdentifierDetector;
-import org.hibernate.tool.internal.export.lint.IssueCollector;
 
 public class HbmLint implements IssueCollector {
